mise a jour du formatage du pdf

+++ src/main/java/com/etat_charge/dto/CalculBalanceDto.java
@@ -2,7 +2,7 @@ package com.etat_charge.dto;
 
 import lombok.*;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 
 @Getter
 @Setter
@@ -14,5 +14,5 @@ public class CalculBalanceDto {
 
     private String accountNumber;
     private String description;
-    private BigDecimal sumDebitBalance;
+    private BigInteger sumDebitBalance;
 }
